notifier/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
plexapi==4.15.9
